+++ src/nes/cpu/opcodes/adc.rs
@@ -31,7 +31,7 @@ impl Adc {
 }
 
 impl FromImmediate for Adc {
-    fn from_immediate(cpu: &mut Cpu) -> usize {
+    fn from_immediate(cpu: &mut Cpu) -> u32 {
         let word = cpu.read_word_and_increment();
         Adc::add_with_carry(cpu, word);
 
@@ -40,7 +40,7 @@ impl FromImmediate for Adc {
 }
 
 impl FromAddress for Adc {
-    fn from_address(cpu: &mut Cpu, mode: AddressMode) -> usize {
+    fn from_address(cpu: &mut Cpu, mode: AddressMode) -> u32 {
         let (src, extra_cycles) = cpu.translate_address(mode);
         let word = cpu.interconnect.read_word(src.to_u16());
 
@@ -49,9 +49,9 @@ impl FromAddress for Adc {
         match mode {
             AddressMode::ZeroPage => 3,
             AddressMode::Absolute => 4,
-            AddressMode::AbsoluteX => { 4 + (extra_cycles as usize) },
-            AddressMode::AbsoluteY => { 4 + (extra_cycles as usize) },
-            AddressMode::IndirectY => { 5 + (extra_cycles as usize) },
+            AddressMode::AbsoluteX => { 4 + (extra_cycles as u32) },
+            AddressMode::AbsoluteY => { 4 + (extra_cycles as u32) },
+            AddressMode::IndirectY => { 5 + (extra_cycles as u32) },
             _ => { panic!("unimplemented address mode {:?} for ADC", mode); }
         }
     }
