+++ src/nes/cpu/opcodes/lda.rs
@@ -11,7 +11,7 @@ impl Lda {
 }
 
 impl FromImmediate for Lda {
-    fn from_immediate(cpu: &mut Cpu) -> usize {
+    fn from_immediate(cpu: &mut Cpu) -> u32 {
         let word = cpu.read_word_and_increment();
         Lda::load_accumulator(cpu, word);
 
@@ -20,7 +20,7 @@ impl FromImmediate for Lda {
 }
 
 impl FromAddress for Lda {
-    fn from_address(cpu: &mut Cpu, mode: AddressMode) -> usize {
+    fn from_address(cpu: &mut Cpu, mode: AddressMode) -> u32 {
         let (src, extra_cycles) = cpu.translate_address(mode);
         let word = cpu.interconnect.read_word(src.to_u16());
 
@@ -30,9 +30,9 @@ impl FromAddress for Lda {
             AddressMode::ZeroPage => 3,
             AddressMode::ZeroPageX => 4,
             AddressMode::Absolute => 4,
-            AddressMode::AbsoluteX => { 4 + (extra_cycles as usize) },
-            AddressMode::AbsoluteY => { 4 + (extra_cycles as usize) },
-            AddressMode::IndirectY => { 5 + (extra_cycles as usize) },
+            AddressMode::AbsoluteX => { 4 + (extra_cycles as u32) },
+            AddressMode::AbsoluteY => { 4 + (extra_cycles as u32) },
+            AddressMode::IndirectY => { 5 + (extra_cycles as u32) },
             _ => { panic!("unimplemented address mode {:?} for LDX", mode); }
         }
     }
