+++ src/nes/cpu/opcodes/ldy.rs
@@ -4,13 +4,13 @@ use super::load::Load;
 pub struct Ldy { }
 
 impl FromImmediate for Ldy {
-    fn from_immediate(cpu: &mut Cpu) -> usize {
-        Load::immediate(cpu, Registers::Y)
+    fn from_immediate(cpu: &mut Cpu) -> u32 {
+        Load::immediate(cpu, Registers::Y) as u32
     }
 }
 
 impl FromAddress for Ldy {
-    fn from_address(cpu: &mut Cpu, mode: AddressMode) -> usize {
+    fn from_address(cpu: &mut Cpu, mode: AddressMode) -> u32 {
         let (src, extra_cycles) = cpu.translate_address(mode);
         let word = cpu.interconnect.read_word(src.to_u16());
         cpu.Y = word;
@@ -20,7 +20,7 @@ impl FromAddress for Ldy {
         match mode {
             AddressMode::ZeroPage => 3,
             AddressMode::ZeroPageX => 4,
-            AddressMode::AbsoluteY => { 4 + (extra_cycles as usize) },
+            AddressMode::AbsoluteY => { 4 + (extra_cycles as u32) },
             _ => { panic!("unimplemented address mode {:?} for LDY", mode); }
         }
     }
